modularise dashboard fiel
