+++ src/features/nextTaskPredictor/components/Popup.tsx
@@ -1,22 +1,7 @@
-import { MemoryRouter as Router, Routes, Route } from 'react-router-dom';
-import HomePage from './HomePage';
-import ResponsePage from './ResponsePage';
-import ChatPage from './ChatPage';
-import DashboardPage from './DashboardPage';
-import AISettingsPage from './AISettingsPage';
+import AppRoutes from '../../../pages/AppRoutes';
 
 const Popup = () => {
-  return (
-    <Router initialEntries={['/']}>
-      <Routes>
-        <Route path="/" element={<HomePage />} />
-        <Route path="/response" element={<ResponsePage />} />
-        <Route path="/chat" element={<ChatPage />} />
-        <Route path="/dashboard" element={<DashboardPage />} />
-        <Route path="/ai-settings" element={<AISettingsPage />} />
-      </Routes>
-    </Router>
-  );
+  return <AppRoutes />;
 };
 
 export default Popup;
